refactor: extract lifecycle handler API into dedicated FormLifecycle.kt
Move registerFormLifecycleHandler, unregisterFormLifecycleHandler, and
KlaviyoFormLifecycleHandlers out of KlaviyoFormsProvider.kt into their
own file, following the pattern of InAppForms.kt in forms-core.

Part of MAGE-287
